optional delete all crd

+++ src/console/API/REST.api.ts
@@ -1,22 +1,10 @@
-import { EMPTY_VALUE_SYMBOL } from '@config/config';
 import { AccessGrantCrdResponse, AccessGrantParams } from '@interfaces/CRD_AccessGrant';
 import { AccessTokenCrdParams, AccessTokenCrdResponse } from '@interfaces/CRD_AccessToken';
-import {
-  CrdStatusCondition,
-  ListCrdResponse,
-  PartialDeploymentResponse,
-  StatusAccessGrantType,
-  StatusAccessTokenType,
-  StatusConnectorType,
-  StatusLinkType,
-  StatusListenerType,
-  StatusSiteType,
-  StatusType
-} from '@interfaces/CRD_Base';
+import { ListCrdResponse, PartialDeploymentResponse } from '@interfaces/CRD_Base';
 import { ConnectorCrdResponse, ConnectorParams } from '@interfaces/CRD_Connector';
 import { LinkCrdResponse } from '@interfaces/CRD_Link';
 import { ListenerCrdParams, ListenerCrdResponse } from '@interfaces/CRD_Listener';
-import { NetworkSite, SiteCrdParams, SiteCrdResponse } from '@interfaces/CRD_Site';
+import { SiteCrdParams, SiteCrdResponse } from '@interfaces/CRD_Site';
 
 import { axiosFetch } from './apiMiddleware';
 import {
@@ -34,7 +22,15 @@ import {
   connectorsPath,
   connectorPath
 } from './REST.paths';
-import { Connector, Listener, Link, SiteView, AccessGrant, StatusAlert } from '../interfaces/REST.interfaces';
+import {
+  convertAccessGrantCRsToAccessGrants,
+  convertConnectorCRsToConnectors,
+  convertLinkCRsToLinks,
+  convertListenerCRsToListeners,
+  convertSiteCRsToSites,
+  getOtherSiteNetworksWithLinks
+} from './REST.utils';
+import { Connector, Listener, Link, SiteView, AccessGrant } from '../interfaces/REST.interfaces';
 
 export const RESTApi = {
   isOldVersionSkupper: async (): Promise<boolean> => {
@@ -59,7 +55,7 @@ export const RESTApi = {
       return null;
     }
 
-    return convertSiteCRsToSite(sites.items[0]);
+    return convertSiteCRsToSites(sites.items[0]);
   },
 
   getSites: async (): Promise<ListCrdResponse<SiteCrdResponse>> =>
@@ -77,25 +73,27 @@ export const RESTApi = {
     return response;
   },
 
-  deleteSite: async (name: string): Promise<void> => {
+  deleteSite: async (name: string, removeAllResources: boolean): Promise<void> => {
     await axiosFetch<SiteCrdResponse>(sitePath(name), {
       method: 'DELETE'
     });
 
-    await Promise.all([
-      axiosFetch<SiteCrdResponse>(accessGrantsPath(), {
-        method: 'DELETE'
-      }),
-      axiosFetch<SiteCrdResponse>(accessTokensPath(), {
-        method: 'DELETE'
-      }),
-      axiosFetch<SiteCrdResponse>(listenersPath(), {
-        method: 'DELETE'
-      }),
-      axiosFetch<SiteCrdResponse>(connectorsPath(), {
-        method: 'DELETE'
-      })
-    ]);
+    if (removeAllResources) {
+      await Promise.all([
+        axiosFetch<SiteCrdResponse>(accessGrantsPath(), {
+          method: 'DELETE'
+        }),
+        axiosFetch<SiteCrdResponse>(accessTokensPath(), {
+          method: 'DELETE'
+        }),
+        axiosFetch<SiteCrdResponse>(listenersPath(), {
+          method: 'DELETE'
+        }),
+        axiosFetch<SiteCrdResponse>(connectorsPath(), {
+          method: 'DELETE'
+        })
+      ]);
+    }
   },
 
   getGrants: async (): Promise<ListCrdResponse<AccessGrantCrdResponse>> =>
@@ -108,7 +106,7 @@ export const RESTApi = {
       return null;
     }
 
-    return convertAccessGrantCRsToAccessGrant(accessGrants.items);
+    return convertAccessGrantCRsToAccessGrants(accessGrants.items);
   },
 
   findGrant: async (name: string): Promise<AccessGrantCrdResponse> =>
@@ -146,6 +144,8 @@ export const RESTApi = {
 
   getLinks: async (): Promise<ListCrdResponse<LinkCrdResponse>> =>
     axiosFetch<ListCrdResponse<LinkCrdResponse>>(linksPath()),
+
+  findLink: async (name: string): Promise<AccessTokenCrdResponse> => axiosFetch<AccessTokenCrdResponse>(linkPath(name)),
 
   getLinksView: async (): Promise<Link[] | null> => {
     const [links, sites] = await Promise.all([RESTApi.getLinks(), RESTApi.getSites()]);
@@ -241,277 +241,3 @@ export const RESTApi = {
     });
   }
 };
-
-// utils
-
-function convertSiteCRsToSite({ metadata, spec, status }: SiteCrdResponse): SiteView {
-  const statusAlertSiteMap: Record<StatusSiteType | 'Error', StatusAlert> = {
-    Configured: 'custom',
-    Running: 'success',
-    Ready: undefined,
-    Resolved: undefined,
-    Error: 'danger'
-  };
-
-  const lastStatus = getLastStatusTrueByTime(status?.conditions);
-  const hasError = lastStatus?.reason === 'Error';
-  const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-
-  const calculatedStatusAlert =
-    (hasError && statusAlertSiteMap.Error) || (lastStatus?.type && statusAlertSiteMap[lastStatus.type]) || undefined;
-
-  return {
-    identity: metadata.uid,
-    name: metadata.name,
-    linkAccess: spec?.linkAccess || '',
-    serviceAccount: spec?.serviceAccount || '',
-    ha: spec?.ha || false,
-    resourceVersion: metadata.resourceVersion,
-    creationTimestamp: metadata.creationTimestamp,
-    linkCount: getOtherSiteNetworksWithLinks(status?.network, metadata.uid).length || 0,
-    isConfigured: hasType(status?.conditions, 'Configured'),
-    isReady: hasType(status?.conditions, 'Ready'),
-    hasError,
-    status: calculatedStatus,
-    platform: findSiteNetwork(status?.network, metadata.uid)?.platform || EMPTY_VALUE_SYMBOL,
-    sitesInNetwork: status?.sitesInNetwork || 0,
-    statusAlert: calculatedStatusAlert
-  };
-}
-
-function convertAccessGrantCRsToAccessGrant(accessGrants: AccessGrantCrdResponse[]): AccessGrant[] {
-  const statusAlertAccessGrantMap: Record<StatusAccessGrantType | 'Error', StatusAlert> = {
-    Processed: 'success',
-    Ready: 'success',
-    Resolved: 'success',
-    Error: 'danger'
-  };
-
-  return accessGrants.map((accessGrant) => {
-    const { metadata, spec, status } = accessGrant;
-
-    const lastStatus = getLastStatusTrueByTime(status?.conditions);
-    const hasError = lastStatus?.reason === 'Error';
-    const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-
-    const calculatedStatusAlert =
-      (hasError && statusAlertAccessGrantMap.Error) ||
-      (lastStatus?.type && statusAlertAccessGrantMap[lastStatus.type]) ||
-      undefined;
-
-    return {
-      id: metadata?.uid,
-      name: metadata?.name,
-      creationTimestamp: metadata?.creationTimestamp,
-      data: accessGrant,
-      hasError,
-      status: calculatedStatus,
-      redemptionsAllowed: spec?.redemptionsAllowed || 0,
-      redeemed: status?.redeemed || 0,
-      expirationWindow: status?.expiration,
-      statusAlert: calculatedStatusAlert
-    };
-  });
-}
-
-function convertLinkCRsToLinks(links: LinkCrdResponse[]): Link[] {
-  const statusAlertLinkMap: Record<StatusLinkType | 'Error', StatusAlert> = {
-    Configured: 'custom',
-    Ready: 'success',
-    Operational: 'success',
-    Error: 'danger'
-  };
-
-  return links.map(({ metadata, spec, status }) => {
-    const lastStatus = getLastStatusTrueByTime(status?.conditions);
-    const hasError = lastStatus?.reason === 'Error';
-    const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-
-    const calculatedStatusAlert =
-      (hasError && statusAlertLinkMap.Error) || (lastStatus?.type && statusAlertLinkMap[lastStatus.type]) || undefined;
-
-    return {
-      id: metadata.uid,
-      name: metadata.name,
-      creationTimestamp: metadata.creationTimestamp,
-      // TODO: The cost needs to be enabled in Skupper v2
-      cost: spec.cost || EMPTY_VALUE_SYMBOL,
-      hasError,
-      status: calculatedStatus,
-      connectedTo: status?.remoteSiteName || EMPTY_VALUE_SYMBOL,
-      statusAlert: calculatedStatusAlert
-    };
-  });
-}
-
-export function convertAccessTokensToLinks(links: AccessTokenCrdResponse[]): Link[] {
-  const statusAlertAccessTokenMap: Record<StatusAccessTokenType | 'Error', StatusAlert> = {
-    Redeemed: 'success',
-    Error: 'danger'
-  };
-
-  return links.map(({ metadata, status }) => {
-    const lastStatus = getLastStatusTrueByTime(status?.conditions);
-    const hasError = lastStatus?.reason === 'Error';
-    const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-    const calculatedStatusAlert =
-      (hasError && statusAlertAccessTokenMap.Error) ||
-      (lastStatus?.type && statusAlertAccessTokenMap[lastStatus.type]) ||
-      undefined;
-
-    return {
-      id: metadata.uid,
-      name: metadata.name,
-      creationTimestamp: metadata.creationTimestamp,
-      cost: EMPTY_VALUE_SYMBOL,
-      hasError,
-      status: calculatedStatus,
-      connectedTo: EMPTY_VALUE_SYMBOL,
-      statusAlert: calculatedStatusAlert
-    };
-  });
-}
-
-function convertListenerCRsToListeners(_: SiteCrdResponse, listeners: ListenerCrdResponse[]): Listener[] {
-  const statusAlertListenerMap: Record<StatusListenerType | 'Error', StatusAlert> = {
-    Configured: 'custom',
-    Matched: 'success',
-    Ready: 'success',
-    Error: 'danger'
-  };
-
-  return listeners.map(({ metadata, spec, status }) => {
-    const lastStatus = getLastStatusTrueByTime(status?.conditions);
-    const hasError = lastStatus?.reason === 'Error';
-    const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-
-    const calculatedStatusAlert =
-      (hasError && statusAlertListenerMap.Error) ||
-      (lastStatus?.type && statusAlertListenerMap[lastStatus.type]) ||
-      undefined;
-
-    return {
-      id: metadata.uid,
-      name: metadata.name,
-      creationTimestamp: metadata.creationTimestamp,
-      routingKey: spec.routingKey,
-      serviceName: spec.host,
-      port: spec.port,
-      type: spec.type,
-      connected: status?.matchingConnectorCount || 0,
-      hasError,
-      status: calculatedStatus,
-      statusAlert: calculatedStatusAlert
-    };
-  });
-}
-
-function convertConnectorCRsToConnectors(_: SiteCrdResponse, connectors: ConnectorCrdResponse[]): Connector[] {
-  const statusAlertConnectorMap: Record<StatusConnectorType | 'Error', StatusAlert> = {
-    Configured: 'custom',
-    Matched: 'success',
-    Ready: 'success',
-    Error: 'danger'
-  };
-
-  return connectors.map(({ metadata, spec, status }) => {
-    const lastStatus = getLastStatusTrueByTime(status?.conditions);
-    const hasError = lastStatus?.reason === 'Error';
-    const calculatedStatus = hasError ? lastStatus.message : lastStatus?.type;
-
-    const calculatedStatusAlert =
-      (hasError && statusAlertConnectorMap.Error) ||
-      (lastStatus?.type && statusAlertConnectorMap[lastStatus.type]) ||
-      undefined;
-
-    return {
-      id: metadata.uid,
-      name: metadata.name,
-      creationTimestamp: metadata.creationTimestamp,
-      routingKey: spec.routingKey,
-      selector: spec.selector,
-      host: spec.host,
-      port: spec.port,
-      type: spec.type,
-      connected: status?.matchingListenerCount || 0,
-      hasError,
-      status: calculatedStatus,
-      statusAlert: calculatedStatusAlert
-    };
-  });
-}
-
-// function findServices(
-//   site: SiteCrdResponse,
-//   routingKey: string,
-//   type: 'listeners' | 'connectors'
-// ): string[] | undefined {
-//   return site.status?.network
-//     ?.filter(({ name, services }) => name !== site.metadata.namespace && services)
-//     .flatMap((network) => network?.services)
-//     .filter((service) => service?.routingKey === routingKey && service[type]?.length)
-//     .flatMap((service) => service![type]) as string[];
-// }
-
-// function isReady(site: SiteCrdResponse, name: string, routingKey: string, type: 'listeners' | 'connectors'): boolean {
-//   return !!site.status?.network
-//     ?.filter((data) => data.name !== site.metadata.namespace && data.services)
-//     .flatMap((network) => network?.services)
-//     .filter((service) => service && service.routingKey === routingKey && service[type]?.includes(name)).length;
-// }
-
-export function getType<T>(conditions: CrdStatusCondition<T>[] = [], type: StatusSiteType | StatusAccessTokenType) {
-  return !!conditions?.some((condition) => condition.type === type && condition.status === 'True');
-}
-
-export function hasType<T>(conditions: CrdStatusCondition<T>[] = [], type: StatusSiteType | StatusAccessTokenType) {
-  return !!conditions?.some((condition) => condition.type === type && condition.status === 'True');
-}
-
-export function hasReasonError<T>(conditions: CrdStatusCondition<T>[] = []) {
-  return !!conditions?.some((condition) => condition.reason === 'Error' && condition.status === 'True');
-}
-
-function getLastStatusTrueByTime<T>(conditions: CrdStatusCondition<T>[] = []) {
-  const priorityMap: Record<StatusType, number> = {
-    Configured: 1,
-    Resolved: 0,
-    Running: 3,
-    Processed: 3,
-    Operational: 3,
-    Matched: 3,
-    Ready: 0,
-    Redeemed: 1
-  };
-
-  let trueConditions = conditions.filter((condition) => condition.status === 'True');
-
-  if (!trueConditions.length) {
-    trueConditions = conditions.filter((condition) => condition.status === 'False' && condition.reason === 'Error');
-  }
-
-  // Sort the filtered conditions by lastTransitionTime in descending order
-  trueConditions.sort((a, b) => {
-    const dateA = new Date(a.lastTransitionTime);
-    const dateB = new Date(b.lastTransitionTime);
-
-    const value = Number(dateB) - Number(dateA);
-
-    if (value === 0) {
-      return priorityMap[b.type as StatusType] - priorityMap[a.type as StatusType];
-    }
-
-    return value;
-  });
-
-  // Return the first condition in the sorted array (the most recent one)
-  return trueConditions.length > 0 ? trueConditions[0] : null;
-}
-
-export function getOtherSiteNetworksWithLinks(network: NetworkSite[] = [], siteId: string) {
-  return network?.filter(({ id, links }) => id !== siteId && links?.some((link) => link.remoteSiteId === siteId));
-}
-
-export function findSiteNetwork(network: NetworkSite[] = [], siteId: string) {
-  return network?.find(({ id }) => id === siteId);
-}
